+++ scripts/dev-server-env.sh
@@ -29,10 +29,19 @@ else
 fi
 
 DEFAULT_DETACH="no"
-if [ -z "$DETACH" ]; then
+if [ -z "${DETACH}" ]; then
   echo "'DETACH' environment variable not set - default of '${DEFAULT_DETACH}' will be used"
   export DC_DETACH=${DEFAULT_DETACH}
 else
-  echo "'DETACH' environment variable set to '$DETACH'"
-  export DC_DETACH=$DETACH
+  echo "'DETACH' environment variable set to '${DETACH}'"
+  export DC_DETACH=${DETACH}
+fi
+
+DEFAULT_ARANGO_VERSION="3.5.3"
+if [ -z "${ARANGO_VERSION}" ]; then
+  echo "'ARANGO_VERSION' environment variable not set - default of '${DEFAULT_ARANGO_VERSION}' will be used"
+  export DC_ARANGO_VERSION=${DEFAULT_ARANGO_VERSION}
+else
+  echo "'ARANGO_VERSION' environment variable is set to '${ARANGO_VERSION}'"
+  export DC_ARANGO_VERSION=${ARANGO_VERSION}
 fi