+++ scripts/stop-dev-server-cluster.sh
@@ -1,0 +1,2 @@
+docker compose -f dev/docker-compose-with-arango-cluster.yml stop
+docker compose  -f dev/docker-compose-with-arango-cluster.yml rm -f